:recycle: Rename and re-organize AccountHome component
