+++ src/modules/Accounts/components/LockedBalanceDetails/index.js
@@ -13,7 +13,7 @@ import ProgressSvg from 'assets/svgs/ProgressSvg';
 import UnlockSvg from 'assets/svgs/UnlockSvg';
 import FormattedNumber from 'components/shared/formattedNumber';
 import getStyles from './styles';
-import { getPendingTime } from '../../utils';
+import { getPendingTime } from '../../utils/accounts.utils';
 import { useAccountInfo } from '../../hooks/useAccountInfo';
 
 const RowItem = ({ styles, title, value, IconComponent, tokenType, language, theme }) => (
