--- src/modules/Accounts/utils/index.js
@@ -1,8 +1,0 @@
-import moment from 'moment';
-
-export const getPendingTime = (unvoteHeight, unlockHeight) => {
-  const awaitingBlocks = unlockHeight - unvoteHeight;
-  const secondsToUnlockAllBalance = awaitingBlocks * 10;
-  const momentSeconds = moment().second(secondsToUnlockAllBalance);
-  return moment().to(momentSeconds, true);
-};
